==> notebooks/1_prep.ipynb <==
# Data Preparation and Exploration

## Setup

#import libraries
import pandas as pd
import numpy as np
import janitor 

First, I will read in my Parquet files.

I will clean the variable names to make them uniform and join the two datasets together.

Then I will remove true duplicates, as each instance should have a unique identifier; according to the metadata (City of Los Angeles, 2025a; 2025b), this is `dr_no` - Division of Records number. Any row that is a complete duplicate is therefore likely to be attributable to a data entry error. 

#get 2010-19 data from parquet
df1 = pd.read_parquet("../data/la_crimes_2010-19.parquet")

#get 2020-25 data from parquet
df2 = pd.read_parquet("../data/la_crimes_2020-25.parquet")

#clean variable names
df1 = (
    df1.clean_names()
    .rename(columns={"date_occ":"date", "time_occ":"time_str", "area_name":"area", "crm_cd":"crime_code", "crm_cd_desc":"crime_type", "premis_cd":"premises_code", "premis_desc":"premises_type", "weapon_used_cd":"weapon_code", "weapon_desc":"weapon_type"})
    )

df2 = (
    df2.clean_names()
    .rename(columns={"date_occ":"date", "time_occ":"time_str", "area":"area_", "area_name":"area", "crm_cd":"crime_code", "crm_cd_desc":"crime_type", "premis_cd":"premises_code", "premis_desc":"premises_type", "weapon_used_cd":"weapon_code", "weapon_desc":"weapon_type"})
    )

#join dataframes and view all columns
df = pd.concat([df1, df2], ignore_index=True)
pd.set_option('display.max_columns', None)
df.head(4)

#remove duplicate rows and check size
df = df.drop_duplicates()
df.shape

## Exploring and Cleaning Variables

### Demographics

I will view the unique values that appear in the demographic variables to check whether they need cleaning, using the metadata to support my understanding and decisions.

#get demographic values and counts
vict_sex_values = df['vict_sex'].value_counts(dropna=False).to_dict()
vict_descent_values = df['vict_descent'].value_counts(dropna=False).to_dict()

print(f"Victim Sex Values:\n{vict_sex_values}\n\nVictim Descent Values:\n{vict_descent_values}")

#tidy victim sex variable
vict_sex_map = {
    "M": "Male",
    "F": "Female",
    "X": "Other/Unknown",
    "H": "Other/Unknown",
    "N": "Other/Unknown",
    "-": "Other/Unknown"
}
df["vict_sex"] = df["vict_sex"].map(vict_sex_map).fillna("Other/Unknown")

#tidy victim descent variable
vict_descent_map = {
    "A": "Other Asian",
    "B": "Black",
    "C": "Chinese",
    "D": "Cambodian",
    "F": "Filipino",
    "G": "Guamanian",
    "H": "Hispanic/Latin/Mexican",
    "I": "American Indian/Alaskan Native",
    "J": "Japanese",
    "K": "Korean",
    "L": "Laotian",
    "O": "Other",
    "P": "Pacific Islander",
    "S": "Samoan",
    "U": "Hawaiian",
    "V": "Vietnamese",
    "W": "White",
    "X": "Unknown",
    "Z": "Asian Indian",
    "-": "Unknown"
}
df["vict_descent"] = df["vict_descent"].map(vict_descent_map).fillna("Unknown")

### Dates and Times

I will clean the date and time variables by converting them into datetime objects. This will allow me to extract granular features later, including day of the week and hour of the day.


#convert dates
df["date"] = pd.to_datetime(df["date"], format="%m/%d/%Y %I:%M:%S %p").dt.normalize()

#convert times
df["time"] = pd.to_datetime(df["time_str"], format="%H%M", errors="coerce").dt.time

#get datetime column
df["datetime_str"] = df["date"].dt.strftime("%Y-%m-%d") + " " + df["time_str"].str[:2] + ":" + df["time_str"].str[2:]
df["datetime"] = pd.to_datetime(df["datetime_str"], format="%Y-%m-%d %H:%M")
df.drop(columns="datetime_str", inplace=True)

### Removing Unnecessary Data

I will drop rows with missing victim age, as this variable is essential for building my model. I will also drop rows where the age is zero or less (vict_age contains many 0s and negative numbers, possibly as crimes without known/human victims e.g. vandalism). Also, I've identified that there are three observations where `vict_age` is unusually high (114, 118, 120). These aren't necessarily entirely impossible as far as ages go, but I think they are unlikely enough that I will drop them in case they are errors. This is further justified as all three have 'Unknown' for `vict_sex` which makes me think that they were related to a non-human victim or somehow logged incorrectly.

#drop columns that won't be used for the model
df = df.drop(columns=["dr_no", "date_rptd", "area_", "rpt_dist_no", "part_1_2", "crime_code", "mocodes", "premises_code", "weapon_code", "status", "status_desc", "crm_cd_1", "crm_cd_2", "crm_cd_3", "crm_cd_4", "location", "cross_street"])

#drop rows with missing victim age
df = df.dropna(subset=["vict_age"])

#get realistic age range (1 to 99)
df = df[df["vict_age"].between(1, 99)]

df.head(4)

## Dataframe Export

# df.to_csv("../data/df_for_grouping.csv", index=False, encoding="utf-8")
