==> notebooks/6b_bayes_cv.ipynb <==
# Final Model Selection - Naïve Bayes

## Setup

#import libraries
import numpy as np
import pandas as pd
from sklearn.model_selection import StratifiedKFold, RandomizedSearchCV, train_test_split
from sklearn.naive_bayes import GaussianNB
from sklearn.metrics import f1_score, recall_score, roc_auc_score, make_scorer, classification_report
from imblearn.over_sampling import SMOTE
from imblearn.pipeline import Pipeline

df = pd.read_csv("../data/df_final.csv")

## Randomized Search Cross-Validation

I will now use `RandomizedSearchCV` to find the best hyperparameters for my Naïve Bayes model. I will use `StratifiedKFold` for cross-validation and `SMOTE` for oversampling, to account for the imbalance in the dataset.

#get features and target
X = df.drop("is_child", axis=1)
y = df["is_child"]

#make pipeline
pipe = Pipeline([
    ("smote", SMOTE(random_state=42)),
    ("clf", GaussianNB())
])

#set hyperparameters
params = {
        "clf__var_smoothing": np.logspace(-11, -7, 10), 
        "smote__k_neighbors": [3, 5, 7],
        "smote__sampling_strategy": [0.6, 0.7, 0.8, 0.9, "auto"],
         }

#instantiate folds for cross-validation
cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

#set scoring metrics
scoring = {
    "recall": make_scorer(recall_score),
    "f1": make_scorer(f1_score),
    "roc_auc": make_scorer(roc_auc_score)
}

#instantiate cross-validation
random_search = RandomizedSearchCV(
    pipe,
    param_distributions=params,
    n_iter=25,
    cv=cv,
    scoring=scoring,
    n_jobs=-1,
    random_state=42,
    refit="recall",
    verbose=2
)

#split into train and test data
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

#fit model
random_search.fit(X_train, y_train)

#get predictions
y_pred = random_search.predict(X_test)

#find best parameters from cross-validation
print("Best parameters:", random_search.best_params_)
print("Best recall score:", random_search.best_score_)

#classification report and ROC AUC
print(classification_report(y_test, y_pred))
print(f"ROC AUC: {roc_auc_score(y_test, y_pred):.2f}")
